feat: implement `useEventsWithTimezone` hook
